app/requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.7.1
Django==3.2
django-tastypie==0.14.6
flup6==1.1.1
future==0.16.0
idna==2.6
mysqlclient==2.1.0
oauthlib==3.2.2
pycryptodomex==3.4.7
pyjwkest==1.4.0
PyJWT==2.8.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
urllib3==2.1.0
uWSGI==2.0.23
